% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdlbrgdown-package.R
\docType{package}
\name{hdlbrgdown-package}
\alias{hdlbrgdown}
\alias{hdlbrgdown-package}
\title{hdlbrgdown: A package for creating PhD theses at the Faculty of Bioscience at 
    Heidelberg University using R Markdown}
\description{
Using the bookdown package and LaTeX and Word thesis templates from
    Reed College, this package aims to encourage reproducible research using
    simple Markdown syntax while embedding all of the R code to produce plots
    and analyses as well.
}
\section{thesis_gitbook}{

Creates an R Markdown thesis template as a webpage
}

\section{thesis_pdf}{

Creates an R Markdown thesis template as a PDF
}

\section{thesis_word}{

Creates an R Markdown thesis template as a Microsoft Word document
}

\section{thesis_epub}{

Creates an R Markdown thesis template as an ebook
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/christianholland/hdlbrgdown}
  \item Report bugs at \url{https://github.com/christianholland/hdlbrgdown/issues}
}

}
\author{
\strong{Maintainer}: Christian H. Holland \email{cholland2408@gmail.com} (\href{https://orcid.org/0000-0002-3060-5786}{ORCID})

}
\keyword{internal}
